+++ lib/queries/two/filter_two_tables.ts
@@ -33,24 +33,20 @@ export class FilterTwoTables<T1, T2> {
         private readonly statement: SelectStatement,
         private readonly filters: number) {}
 
-    filter(predicate: (first: T1, second: T2) => boolean): FilterTwoTables<T1, T2> {
-        return new FilterTwoTables(
-            this.firstConstructor,
-            this.secondConstructor,
-            {
-                ...this.statement,
-                filters: this.statement.filters.concat(parseParameterlessFilter(predicate))
-            },
-            this.filters+1)
-    }
+    filter(predicate: (first: T1, second: T2) => boolean): FilterTwoTables<T1, T2>
+    filter<P extends ValueOrNestedValueRecord>(provided: P, predicate: (parameters: P, first: T1, second: T2) => boolean): FilterTwoTables<T1, T2>
+    filter<P extends ValueOrNestedValueRecord>(predicateOrProvided: ((first: T1, second: T2) => boolean)|P, predicate?: (parameters: P, first: T1, second: T2) => boolean): FilterTwoTables<T1, T2> {
 
-    filterP<P extends ValueOrNestedValueRecord>(provided: P, predicate: (parameter: P, first: T1, second: T2) => boolean): FilterTwoTables<T1, T2> {
+        const additionalFilter = typeof predicateOrProvided === 'function'
+            ? parseParameterlessFilter(predicateOrProvided)
+            : parseParameterizedFilter(predicate!, `f${this.filters + 1}`, predicateOrProvided)
+
         return new FilterTwoTables(
             this.firstConstructor,
             this.secondConstructor,
             {
                 ...this.statement,
-                filters: this.statement.filters.concat(parseParameterizedFilter(predicate, `f${this.filters + 1}`, provided))
+                filters: this.statement.filters.concat(additionalFilter)
             },
             this.filters+1)
     }
@@ -86,21 +82,17 @@ export class FilterTwoTables<T1, T2> {
             })
     }
 
-    map<U extends ValueRecord>(f: (first: T1, second: T2) => EnforceNonEmptyRecord<U> & U): SelectRows<U> {
-        return new SelectRows(
-            {
-                ...this.statement,
-                selection: parseMapSelection(f)
-            })
-    }
+    map<U extends ValueRecord>(f: (first: T1, second: T2) => EnforceNonEmptyRecord<U> & U): SelectRows<U>
+    map<S, U extends ValueRecord>(tableInSubquery: Table<S>, f: (s: Subtable<S>, first: T1, second: T2) => EnforceNonEmptyRecord<U> & U): SelectRows<U>
+    map<S, U extends ValueRecord>(fOrTableInSubquery: ((first: T1, second: T2) => EnforceNonEmptyRecord<U> & U)|Table<S>, f?: (s: Subtable<S>, first: T1, second: T2) => EnforceNonEmptyRecord<U> & U): SelectRows<U>{
+        const selection = typeof fOrTableInSubquery === 'function'
+            ? parseMapSelection(fOrTableInSubquery)
+            : parseMapWithSubquerySelection(f!, [fOrTableInSubquery.tableName])
 
-    mapS<S, U extends ValueRecord>(
-        tableInSubquery: Table<S>,
-        f: (s: Subtable<S>, first: T1, second: T2) => EnforceNonEmptyRecord<U> & U): SelectRows<U> {
         return new SelectRows(
             {
                 ...this.statement,
-                selection: parseMapWithSubquerySelection(f, [tableInSubquery.tableName])
+                selection
             })
     }